djangoapp/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
djoser==2.2.0
djangorestframework==3.14.0
python-dotenv>=1.0.0,<1.1
django-axes>=6.0.1,<6.1
